split ClusteringWriter tests
