+++ eva-accession-clustering/src/test/java/uk/ac/ebi/eva/accession/clustering/batch/io/clustering_writer/MergeAccessionClusteringWriterTest.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.clustering.batch.io;
+package uk.ac.ebi.eva.accession.clustering.batch.io.clustering_writer;
 
-import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbConfigurationBuilder;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbRule;
 import com.mongodb.client.FindIterable;
@@ -36,15 +35,12 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.ac.ebi.ampt2d.commons.accession.core.exceptions.AccessionCouldNotBeGeneratedException;
-import uk.ac.ebi.ampt2d.commons.accession.core.exceptions.AccessionDeprecatedException;
-import uk.ac.ebi.ampt2d.commons.accession.core.exceptions.AccessionDoesNotExistException;
-import uk.ac.ebi.ampt2d.commons.accession.core.exceptions.AccessionMergedException;
 import uk.ac.ebi.ampt2d.commons.accession.core.models.EventType;
 import uk.ac.ebi.ampt2d.commons.accession.hashing.SHA1HashingFunction;
 import uk.ac.ebi.ampt2d.commons.accession.persistence.jpa.monotonic.repositories.ContiguousIdBlockRepository;
 import uk.ac.ebi.ampt2d.commons.accession.persistence.mongodb.document.EventDocument;
 
+import uk.ac.ebi.eva.accession.clustering.batch.io.ClusteringWriter;
 import uk.ac.ebi.eva.accession.clustering.parameters.InputParameters;
 import uk.ac.ebi.eva.accession.clustering.test.configuration.BatchTestConfiguration;
 import uk.ac.ebi.eva.accession.clustering.test.rule.FixSpringMongoDbRule;
@@ -68,25 +64,28 @@ import uk.ac.ebi.eva.accession.core.summary.ClusteredVariantSummaryFunction;
 import uk.ac.ebi.eva.accession.core.summary.SubmittedVariantSummaryFunction;
 import uk.ac.ebi.eva.commons.core.models.VariantType;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.function.Function;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static uk.ac.ebi.eva.accession.clustering.test.VariantAssertions.assertAccessionEqual;
 import static uk.ac.ebi.eva.accession.clustering.test.VariantAssertions.assertAssemblyAccessionEqual;
 import static uk.ac.ebi.eva.accession.clustering.test.VariantAssertions.assertClusteredVariantAccessionEqual;
 import static uk.ac.ebi.eva.accession.clustering.test.VariantAssertions.assertReferenceSequenceAccessionEqual;
 
+/**
+ * This class handles some scenarios of ClusteringWriter where redundant RSs are discovered and merged.
+ *
+ * That includes all the scenarios that involve writing to the Operation collections, e.g.
+ * clusteredVariantOperationEntity to register an RS merge.
+ * Other test classes in this folder take care of other scenarios.
+ */
 @RunWith(SpringRunner.class)
 @EnableAutoConfiguration
 @ContextConfiguration(classes = {ClusteredVariantAccessioningConfiguration.class, BatchTestConfiguration.class})
 @TestPropertySource("classpath:clustering-pipeline-test.properties")
-public class ClusteringWriterTest {
+public class MergeAccessionClusteringWriterTest {
 
     private static final String TEST_DB = "test-db";
 
@@ -97,8 +96,6 @@ public class ClusteringWriterTest {
     public static final String DBSNP_CLUSTERED_VARIANT_COLLECTION = "dbsnpClusteredVariantEntity";
 
     public static final String DBSNP_SUBMITTED_VARIANT_COLLECTION = "dbsnpSubmittedVariantEntity";
-
-    private static final String PROJECT_ACCESSION = "projectId_1";
 
     public static final long EVA_CLUSTERED_VARIANT_RANGE_START = 3000000000L;
 
@@ -148,176 +145,6 @@ public class ClusteringWriterTest {
         mongoTemplate.dropCollection(ClusteredVariantOperationEntity.class);
         mongoTemplate.dropCollection(DbsnpSubmittedVariantOperationEntity.class);
         mongoTemplate.dropCollection(DbsnpClusteredVariantOperationEntity.class);
-    }
-
-    @Test
-    @UsingDataSet(locations = {"/test-data/submittedVariantEntity.json"})
-    @DirtiesContext
-    public void writer() throws Exception {
-        List<SubmittedVariantEntity> submittedVariantEntities = createSubmittedVariantEntities();
-        clusteringWriter.write(submittedVariantEntities);
-        assertClusteredVariantsCreated();
-        assertSubmittedVariantsUpdated();
-    }
-
-    private List<SubmittedVariantEntity> createSubmittedVariantEntities() {
-        List<SubmittedVariantEntity> submittedVariantEntities = new ArrayList<>();
-        SubmittedVariant submittedVariant1 = createSubmittedVariant(inputParameters.getAssemblyAccession(), 1000,
-                                                                    PROJECT_ACCESSION, "1", 1000L, "T", "A");
-        SubmittedVariantEntity submittedVariantEntity1 = createSubmittedVariantEntity(1L, submittedVariant1);
-        //Different alleles
-        SubmittedVariant submittedVariant2 = createSubmittedVariant(inputParameters.getAssemblyAccession(), 1000,
-                                                                    PROJECT_ACCESSION, "1", 1000L, "T", "G");
-        SubmittedVariantEntity submittedVariantEntity2 = createSubmittedVariantEntity(2L, submittedVariant2);
-        //Same assembly, contig, start but different type
-        SubmittedVariant submittedVariantINS = createSubmittedVariant(inputParameters.getAssemblyAccession(), 1000,
-                                                                      PROJECT_ACCESSION, "1", 1000L, "", "A");
-        SubmittedVariantEntity submittedVariantEntityINS = createSubmittedVariantEntity(3L, submittedVariantINS);
-        SubmittedVariant submittedVariantDEL = createSubmittedVariant(inputParameters.getAssemblyAccession(), 1000,
-                                                                      PROJECT_ACCESSION, "1", 1000L, "T", "");
-        SubmittedVariantEntity submittedVariantEntityDEL = createSubmittedVariantEntity(4L, submittedVariantDEL);
-        //Different assembly, contig and start
-        SubmittedVariant submittedVariant3 = createSubmittedVariant(inputParameters.getAssemblyAccession(), 3000,
-                                                                    PROJECT_ACCESSION, "1", 3000L, "C", "G");
-        SubmittedVariantEntity submittedVariantEntity3 = createSubmittedVariantEntity(5L, submittedVariant3);
-        submittedVariantEntities.add(submittedVariantEntity1);
-        submittedVariantEntities.add(submittedVariantEntity2);
-        submittedVariantEntities.add(submittedVariantEntityINS);
-        submittedVariantEntities.add(submittedVariantEntityDEL);
-        submittedVariantEntities.add(submittedVariantEntity3);
-        return submittedVariantEntities;
-    }
-
-    private SubmittedVariantEntity createSubmittedVariantEntity(Long accession, SubmittedVariant submittedVariant) {
-        String hash = hashingFunction.apply(submittedVariant);
-        SubmittedVariantEntity submittedVariantEntity = new SubmittedVariantEntity(accession, hash, submittedVariant, 1);
-        return submittedVariantEntity;
-    }
-
-    private SubmittedVariant createSubmittedVariant(String referenceSequenceAccession, int taxonomyAccession,
-                                                    String projectAccession, String contig, long start,
-                                                    String referenceAllele, String alternateAllele) {
-        return new SubmittedVariant(referenceSequenceAccession, taxonomyAccession, projectAccession, contig, start,
-                                    referenceAllele, alternateAllele, null);
-    }
-
-    private void assertClusteredVariantsCreated() {
-        MongoCollection<Document> collection = mongoTemplate.getCollection(CLUSTERED_VARIANT_COLLECTION);
-        assertEquals(4, collection.countDocuments());
-        List<Long> expectedAccessions = Arrays.asList(3000000000L, 3000000001L, 3000000002L, 3000000003L);
-        assertGeneratedAccessions(CLUSTERED_VARIANT_COLLECTION, "accession", expectedAccessions);
-    }
-
-    private void assertGeneratedAccessions(String collectionName, String accessionField,
-                                           List<Long> expectedAccessions) {
-        List<Long> generatedAccessions = new ArrayList<>();
-        MongoCollection<Document> collection = mongoTemplate.getCollection(collectionName);
-        FindIterable<Document> dbObjects = collection.find();
-        for (Document document : dbObjects) {
-            Long accessionId = (Long) document.get(accessionField);
-            generatedAccessions.add(accessionId);
-        }
-        Collections.sort(generatedAccessions);
-        assertEquals(expectedAccessions, generatedAccessions);
-    }
-
-    private void assertSubmittedVariantsUpdated() {
-        assertTrue(allSubmittedVariantsClustered());
-        List<Long> expectedAccessions = Arrays.asList(3000000000L, 3000000000L, 3000000001L, 3000000002L, 3000000003L);
-        assertGeneratedAccessions(SUBMITTED_VARIANT_COLLECTION, "rs", expectedAccessions);
-    }
-
-    private boolean allSubmittedVariantsClustered() {
-        MongoCollection<Document> collection = mongoTemplate.getCollection(SUBMITTED_VARIANT_COLLECTION);
-        FindIterable<Document> documents = collection.find();
-        for (Document document : documents) {
-            if (document.get("rs") == null){
-                return false;
-            }
-        }
-        return true;
-    }
-
-    @Test
-    @DirtiesContext
-    public void reuse_clustered_accession_if_provided() throws AccessionCouldNotBeGeneratedException, AccessionMergedException, AccessionDoesNotExistException, AccessionDeprecatedException {
-        long existingRs = 30L;
-        String asm1 = "asm1";
-        String asm2 = "asm2";
-
-        ClusteredVariantEntity cve1 = createClusteredVariantEntity(asm1, existingRs);
-        mongoTemplate.insert(cve1);
-        assertEquals(1, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-
-        // given a submitted variant that already has an RS=rs30
-        SubmittedVariantEntity sveClustered = createSubmittedVariantEntity(asm1, existingRs, 50L);
-        clusteringWriter.write(Collections.singletonList(sveClustered));
-        assertEquals(1, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-
-        // given a different submitted variant without an assigned RS (rs=null)
-        SubmittedVariantEntity sveNonClustered = createSubmittedVariantEntity(asm2, null, 51L);
-        clusteringWriter.write(Collections.singletonList(sveNonClustered));
-        assertEquals(2, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-
-        // given the same submitted variant without an assigned RS (rs=null), getOrCreate should not create another RS
-        clusteringWriter.write(Collections.singletonList(sveNonClustered));
-        assertEquals(2, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-    }
-
-    private SubmittedVariantEntity createSubmittedVariantEntity(String assembly, Long rs, Long ss) {
-        SubmittedVariant submittedClustered = new SubmittedVariant(assembly, 1000, "project", "1", 100L, "T", "A", rs);
-        String hash1 = hashingFunction.apply(submittedClustered);
-        return new SubmittedVariantEntity(ss, hash1, submittedClustered, 1);
-    }
-
-    private ClusteredVariantEntity createClusteredVariantEntity(String assembly, Long rs) {
-        ClusteredVariant cv = new ClusteredVariant(assembly, 1000, "1", 100L, VariantType.SNV, false, null);
-        String cvHash = clusteredHashingFunction.apply(cv);
-        return new ClusteredVariantEntity(rs, cvHash, cv, 1);
-    }
-
-    private DbsnpClusteredVariantEntity createDbsnpClusteredVariantEntity(String assembly, Long rs) {
-        ClusteredVariant cv = new ClusteredVariant(assembly, 1000, "1", 100L, VariantType.SNV, false, null);
-        String cvHash = clusteredHashingFunction.apply(cv);
-        return new DbsnpClusteredVariantEntity(rs, cvHash, cv, 1);
-    }
-
-    @Test
-    @DirtiesContext
-    public void reuse_dbsnp_clustered_accession_when_clustering_an_eva_submitted_variant() throws Exception {
-        // given
-        Long rs1 = 30L;
-
-        assertEquals(0, mongoTemplate.count(new Query(), DbsnpSubmittedVariantEntity.class));
-        assertEquals(0, mongoTemplate.count(new Query(), DbsnpClusteredVariantEntity.class));
-        assertEquals(0, mongoTemplate.count(new Query(), SubmittedVariantEntity.class));
-        assertEquals(0, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-
-        ClusteredVariant cv1 = new ClusteredVariant("asm1", 1000, "1", 100L, VariantType.SNV, false, null);
-        String cvHash1 = clusteredHashingFunction.apply(cv1);
-        DbsnpClusteredVariantEntity cve1 = new DbsnpClusteredVariantEntity(rs1, cvHash1, cv1, 1);
-        mongoTemplate.insert(cve1);
-
-        SubmittedVariant submittedNonClustered = new SubmittedVariant("asm1", 1000, "project2", "1", 100L, "T", "A",
-                                                                      null);
-        SubmittedVariantEntity sveNonClustered = createSubmittedVariantEntity(51L, submittedNonClustered);
-        mongoTemplate.insert(sveNonClustered);
-
-        assertEquals(0, mongoTemplate.count(new Query(), DbsnpSubmittedVariantEntity.class));
-        assertEquals(1, mongoTemplate.count(new Query(), DbsnpClusteredVariantEntity.class));
-        assertEquals(1, mongoTemplate.count(new Query(), SubmittedVariantEntity.class));
-        assertEquals(0, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-
-        // when
-        clusteringWriter.write(Collections.singletonList(sveNonClustered));
-
-        // then
-        assertEquals(0, mongoTemplate.count(new Query(), DbsnpSubmittedVariantEntity.class));
-        assertEquals(1, mongoTemplate.count(new Query(), DbsnpClusteredVariantEntity.class));
-        assertEquals(1, mongoTemplate.count(new Query(), SubmittedVariantEntity.class));
-        assertEquals(0, mongoTemplate.count(new Query(), ClusteredVariantEntity.class));
-        SubmittedVariantEntity afterClustering = mongoTemplate.findOne(new Query(), SubmittedVariantEntity.class);
-        assertEquals(rs1, afterClustering.getClusteredVariantAccession());
     }
 
     @Test
@@ -375,16 +202,7 @@ public class ClusteringWriterTest {
         assertMergedInto(rs1, rs2, ss2);
     }
 
-    @Test
-    @DirtiesContext
-    public void reuse_eva_clustered_accession_for_a_dbsnp_submitted_variant() throws Exception {
-        Long rs1 = null;
-        Long rs2 = 3100000000L;
-        Long ssToRemap = 50L;
-        Long ss2 = 5100000000L;
-        mergeClusteredAccession(rs1, rs2, ssToRemap, ss2, 0, 1, 0, 0, 0, 2, 0, 1);
-//        assertMergedInto(rs1, rs2, ss2);
-    }
+    // TODO: test that merging is not performed when an RS has several locus in the same assembly
 
     public void mergeClusteredAccession(Long rs1, Long rs2, Long ssToRemap, Long ss2,
                                         int expectedDbsnpCve, int expectedCve,
@@ -431,32 +249,16 @@ public class ClusteringWriterTest {
         assertAssemblyAccessionEqual(Sets.newTreeSet(asm1, asm2), clusteredVariants);
     }
 
-    private void assertMergedInto(Long mergedInto, Long originalAccession, Long updatedSubmittedVariant) {
-        List<SubmittedVariantEntity> submittedVariants = mongoTemplate.findAll(SubmittedVariantEntity.class);
-        submittedVariants.addAll(mongoTemplate.findAll(DbsnpSubmittedVariantEntity.class));
-        assertClusteredVariantAccessionEqual(Sets.newTreeSet(mergedInto), submittedVariants);
+    private ClusteredVariantEntity createClusteredVariantEntity(String assembly, Long rs) {
+        ClusteredVariant cv = new ClusteredVariant(assembly, 1000, "1", 100L, VariantType.SNV, false, null);
+        String cvHash = clusteredHashingFunction.apply(cv);
+        return new ClusteredVariantEntity(rs, cvHash, cv, 1);
+    }
 
-        List<ClusteredVariantEntity> clusteredVariants = mongoTemplate.findAll(ClusteredVariantEntity.class);
-        clusteredVariants.addAll(mongoTemplate.findAll(DbsnpClusteredVariantEntity.class));
-        assertAccessionEqual(Sets.newTreeSet(mergedInto), clusteredVariants);
-
-        EventDocument<IClusteredVariant, Long, ? extends ClusteredVariantInactiveEntity> clusteredOp =
-                mongoTemplate.findOne(new Query(), ClusteredVariantOperationEntity.class);
-        if (clusteredOp == null) {
-            clusteredOp = mongoTemplate.findOne(new Query(), DbsnpClusteredVariantOperationEntity.class);
-        }
-        assertEquals(EventType.MERGED, clusteredOp.getEventType());
-        assertEquals(originalAccession, clusteredOp.getAccession());
-        assertEquals(mergedInto, clusteredOp.getMergedInto());
-
-
-        EventDocument<ISubmittedVariant, Long, ? extends SubmittedVariantInactiveEntity> submittedOp =
-                mongoTemplate.findOne(new Query(), SubmittedVariantOperationEntity.class);
-        if (submittedOp == null) {
-            submittedOp = mongoTemplate.findOne(new Query(), DbsnpSubmittedVariantOperationEntity.class);
-        }
-        assertEquals(EventType.UPDATED, submittedOp.getEventType());
-        assertEquals(updatedSubmittedVariant, submittedOp.getAccession());
+    private SubmittedVariantEntity createSubmittedVariantEntity(String assembly, Long rs, Long ss) {
+        SubmittedVariant submittedClustered = new SubmittedVariant(assembly, 1000, "project", "1", 100L, "T", "A", rs);
+        String hash1 = hashingFunction.apply(submittedClustered);
+        return new SubmittedVariantEntity(ss, hash1, submittedClustered, 1);
     }
 
     private String getClusteredTable(Long rs1) {
@@ -489,5 +291,33 @@ public class ClusteringWriterTest {
         assertEquals(expectedDbsnpSvOperations, mongoTemplate.count(new Query(),
                                                                     DbsnpSubmittedVariantOperationEntity.class));
         assertEquals(expectedSvOperationsint, mongoTemplate.count(new Query(), SubmittedVariantOperationEntity.class));
+    }
+
+    private void assertMergedInto(Long mergedInto, Long originalAccession, Long updatedSubmittedVariant) {
+        List<SubmittedVariantEntity> submittedVariants = mongoTemplate.findAll(SubmittedVariantEntity.class);
+        submittedVariants.addAll(mongoTemplate.findAll(DbsnpSubmittedVariantEntity.class));
+        assertClusteredVariantAccessionEqual(Sets.newTreeSet(mergedInto), submittedVariants);
+
+        List<ClusteredVariantEntity> clusteredVariants = mongoTemplate.findAll(ClusteredVariantEntity.class);
+        clusteredVariants.addAll(mongoTemplate.findAll(DbsnpClusteredVariantEntity.class));
+        assertAccessionEqual(Sets.newTreeSet(mergedInto), clusteredVariants);
+
+        EventDocument<IClusteredVariant, Long, ? extends ClusteredVariantInactiveEntity> clusteredOp =
+                mongoTemplate.findOne(new Query(), ClusteredVariantOperationEntity.class);
+        if (clusteredOp == null) {
+            clusteredOp = mongoTemplate.findOne(new Query(), DbsnpClusteredVariantOperationEntity.class);
+        }
+        assertEquals(EventType.MERGED, clusteredOp.getEventType());
+        assertEquals(originalAccession, clusteredOp.getAccession());
+        assertEquals(mergedInto, clusteredOp.getMergedInto());
+
+
+        EventDocument<ISubmittedVariant, Long, ? extends SubmittedVariantInactiveEntity> submittedOp =
+                mongoTemplate.findOne(new Query(), SubmittedVariantOperationEntity.class);
+        if (submittedOp == null) {
+            submittedOp = mongoTemplate.findOne(new Query(), DbsnpSubmittedVariantOperationEntity.class);
+        }
+        assertEquals(EventType.UPDATED, submittedOp.getEventType());
+        assertEquals(updatedSubmittedVariant, submittedOp.getAccession());
     }
 }
